Added lock/unlock method to the RPC server so the RPC server maintains the robot data mutex itself. The robot controller now have to use the provided API for lock/unlock. Photo sensor Dev RPC call returns double instead of int.

+++ librapirpc/robotrpcclient.cpp
@@ -118,11 +118,11 @@ bool RobotRpcClient::getCliffDev(unsigned int& numSamples)
 }
 
 //------------------------------------------------------------------------------
-bool RobotRpcClient::getPhotoDev(unsigned int& numSamples, unsigned int& maxRange)
+bool RobotRpcClient::getPhotoDev(unsigned int& numSamples, double& maxRange)
 {
     object result = call("getPhotoDev", object());
     numSamples = fromVariant<int>(result["numSamples"]);
-    maxRange = fromVariant<int>(result["maxRange"]);
+    maxRange = fromVariant<double>(result["maxRange"]);
     return true;
 }
